Move pure JS-code for PayloadOutputFunctions into a static script.
I.e. pull it in through Angular's scripts option rather than through a TS import.
Declare corresponding TS types in the PayloadOutputService.

This commit was supposed to exclude PayloadOutputFunctions from minification by the Angular build, so that they can be displayed nicely in the UI.
But Angular still minifies stuff that comes in through the scripts options.
More work needed.